NOTE:  Format specification has changed.
Previously the keyword separator was a single ':'.  However this
conflicted with Fortran's use of ':' as an edit-descriptor.  The
simplest change was to require the separator to be '::', which is not
a valid Fortran edit descriptor.

This functionality is useful for formatting arrays where the separator
between array elements should not be repeated after the last element.
E.g.

str = parser%makeString('%(100(f10.2,:,", "))', arr1D_1=[1.,2.,3.,4.])

The ", " will be inserted between array elements (i.e. 3 times), but
the ':' prevents the 4th occurrence.