A recent bug fix for pFUnit is now correctly identifying cases where test expected strings had an extra " " (space) at the end of lines.
Unfortunately with the current approach, there is no way to correctly
capture such trailing spaces.  A fixed sized buffer is required for
Fortran internal files, which leaves only len() and len_trim() as the
means to determine the length of a given line. 

Because this package provides other means to introduce spaces if
desired, we choose to apply len_trim() and thus ignore trailing spaces
for each position and keyword output in FormatString.