Various changes to enable 'missing' loggers in manager hierarchy.
The Placeholder scheme in Python logging requires that the container
of loggers support objects that have pointers to their parents also in
the container.  The initial implementation of MAP in FTL did not
guarantee persistence of such pointers as the parent might be copied
to another location under further insertions.

An option was therefore added in FTL to allow MAPs to have the
ALLOCATABLE attribute for the pair data structure that is managed.
With this setting pairs are always moved with MOVE_ALLOC instead of
copied.  (This unfortunately breaks gfortran 5.2 for nested maps with
CLASS(*) used in Config.  Otherwise this option would have become the
default.)