Extending support for Formatter for MPI use case.
This work is still evolving, but the framework needs to provide a
means to include rank and npes info in formatted output.  These depend
upon communicators.

Current approach defines a subclass: MpiFormatter that takes comm (or
comms(:)) as arguments.  There are also optional arguments to override
the default keywords of 'rank' and 'npes'.

Next: extend the basic formatter to accept an optional dictionary for
preprocessing part of the fmt string.  Only have an MPI use case for
now, but it is not hard to imagine other such cases.  At that point,
MpiFormatter may no longer be needed as a separate type.