Added another section to the config file for the root logger.
Previously, there was no mechanism to configure the root_node of a
logger manager, which is actually often the only logger to get
handlers.  Mostly just had not gotten around to adding that
functionality.  Trivial extension of existing logic.

Also updated the examples for MPI to use the new feature.