Changed the way to override default communicator.
The optional "extra" parameter is now exclusively used to get
communicator values during configuration.  MPI_COMM_WORLD is added to
extra unless load_file() specifies an alternative communicator.   The global
communicator is referred to as "COMM_LOGGER".

Note that if the user specifies MPI_COMM_WORLD, it will actually use
COMM_LOGGER.   Possibly a warning should be sent at that point, but this
is not currently implemented.