Corrected problem with duplicating locks.
Previously MpiLock was creating MPI windows and such during
construction of objects.  But if we want to copy the lock for use in
multiple handlers then this is overly strict.  We really want a truly
separate lock on each handler.

Now the init does nothing except save the value of the comm.  The real
work of building the object is performed in the init() procedure (and undone
in the destroy procedure).

Untested.