Workaround for gfortran.
pFlogger uses an allocatable array of dynamically sized strings for
intermediate formatting of user-specified formats.  GFortran will only
work with fixed-length strings in this situation.

Until now, the workaround was to used fixed length strings in gfortran
and throw an exception if the string is not long enough.

Other work has introduced the "\n" notation into formats which is an
acceptable alternative for gfortran.

This change deactivates support for multi-record formatns (using slash
edit descriptor.)  That's too far and an intermediate solution is now
desired.