Fix/workraround for mock mpi testing
The use of link-time polymorphism to mock MPI has always been a bit
risky.    It looks lik recent changes in NAG and/or OpenMPI
necessitate some changes.  Some are generally good on their own.  But
there is some risk this is now less portable to other flavors of MPI.

Since the hack only affects tests, we can restrict CI testing to only
be OpenMPI in the worst case.